requirements.txt
asgiref==3.2.10
Django==3.0.8
django-crispy-forms==1.9.2
django-settings-json==1.0.2
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
